Move OpenGL stuff into own package
